[refactor] 폴더링 변경
